move status effects and EntityNetworkevents over to the api.  Fix JavaDoc text on SignComponent
