+++ src/main/java/org/spout/vanilla/api/event/entity/network/EntityRemoveEffectEvent.java
@@ -24,7 +24,7 @@
  * License and see <http://spout.in/licensev1> for the full license, including
  * the MIT license.
  */
-package org.spout.vanilla.plugin.event.entity.network;
+package org.spout.vanilla.api.event.entity.network;
 
 import org.spout.api.entity.Entity;
 import org.spout.api.event.HandlerList;
