+++ src/main/java/org/spout/vanilla/plugin/component/misc/EffectsComponent.java
@@ -37,9 +37,9 @@ import org.spout.vanilla.api.data.effect.StatusEffect;
 
 import org.spout.vanilla.plugin.component.living.Living;
 import org.spout.vanilla.plugin.component.living.neutral.Human;
-import org.spout.vanilla.plugin.data.effect.StatusEffectContainer;
-import org.spout.vanilla.plugin.event.entity.network.EntityEffectEvent;
-import org.spout.vanilla.plugin.event.entity.network.EntityRemoveEffectEvent;
+import org.spout.vanilla.api.data.effect.StatusEffectContainer;
+import org.spout.vanilla.api.event.entity.network.EntityEffectEvent;
+import org.spout.vanilla.api.event.entity.network.EntityRemoveEffectEvent;
 import org.spout.vanilla.plugin.material.item.misc.Potion;
 
 public class EffectsComponent extends EntityComponent {
