+++ src/main/java/org/spout/vanilla/plugin/material/item/Food.java
@@ -40,7 +40,7 @@ import org.spout.vanilla.plugin.component.living.Living;
 import org.spout.vanilla.plugin.component.misc.EffectsComponent;
 import org.spout.vanilla.plugin.component.misc.HungerComponent;
 import org.spout.vanilla.plugin.data.VanillaData;
-import org.spout.vanilla.plugin.data.effect.StatusEffectContainer;
+import org.spout.vanilla.api.data.effect.StatusEffectContainer;
 
 public class Food extends VanillaItemMaterial {
 	private final FoodEffect[] effects;
