+++ src/main/java/org/spout/vanilla/plugin/protocol/VanillaNetworkSynchronizer.java
@@ -91,8 +91,8 @@ import org.spout.vanilla.api.event.entity.EntityCollectItemEvent;
 import org.spout.vanilla.api.event.entity.EntityEquipmentEvent;
 import org.spout.vanilla.api.event.entity.EntityMetaChangeEvent;
 import org.spout.vanilla.api.event.entity.EntityStatusEvent;
-import org.spout.vanilla.plugin.event.entity.network.EntityEffectEvent;
-import org.spout.vanilla.plugin.event.entity.network.EntityRemoveEffectEvent;
+import org.spout.vanilla.api.event.entity.network.EntityEffectEvent;
+import org.spout.vanilla.api.event.entity.network.EntityRemoveEffectEvent;
 import org.spout.vanilla.api.event.item.MapItemUpdateEvent;
 import org.spout.vanilla.plugin.event.player.network.PlayerAbilityUpdateEvent;
 import org.spout.vanilla.api.event.player.network.PlayerBedEvent;
